--- .env
@@ -1,2 +1,0 @@
-PORT=9000
-MONGODB_URL=mongodb://localhost:27017/go-auth
